+++ src/journeyTest/kotlin/de/bund/digitalservice/useid/config/JourneyTestConfig.kt
@@ -1,4 +1,4 @@
-package de.bund.digitalservice.useid.journey
+package de.bund.digitalservice.useid.config
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.boot.test.context.TestConfiguration
